Added two tests. Reorganized variables, refactored, and created new method.
Two additional tests were added, multipleSpacesTest() and longerTextTest().  Variables were moved out of the wrap() method and converted to private class variables so that the new method, calculateWastedSpace(), could have access to them.  Cleaned up comments and tweaked the primary if-statement condition to properly calculate when to add a newline.