Removed calculateWastedSpace, simplified main if-statement condition, added another test case.
A much simpler version of the main if-statement condition was implemented resulting in multiple variables and the calculateWastedSpace method to become unused. Added two new tests which test if words longer than the line length near the end of the text throw exceptions, and another test with multiple spaces in the middle of the text.