Moved variables variables, added test with tab character.
The class variables were moved back in to the wrap method. This was causing the tests to fail when run together. A new test containing a tab character in the input was added. This resulted in two new variables and changes in the if-statement condition that checked if the character count was sufficiently far away from the previous new line to add the next new line.